fixed issue with Play component rendering, added playerGrid logic into respective component and rendered it within Play
