+++ src/App.js
@@ -2,7 +2,7 @@ import './App.css';
 import Header from './components/Header';
 import LandingPage from './components/LandingPage';
 import Footer from './components/Footer'
-// import Play from './Play';
+import Play from './components/Play'
 import GetRocket from './components/GetRocket';
 import { Route, Routes } from 'react-router-dom';
 
@@ -14,7 +14,7 @@ function App() {
         <Routes>
           <Route path = "/" element={<LandingPage />} />
           <Route path="/rocket-selection" element={<GetRocket />} />
-          {/* <Route path="/play" element={<Play />} /> */}
+          <Route path="/play" element={<Play />} />
         </Routes> 
       </main>
       {/* <Footer /> */}
