+++ src/components/Play.js
@@ -1,11 +1,14 @@
 // PLAY Component
-import { GetRocket, Score, PlayerGrid, NPCGrid} from '../components'
+import GetRocket from "./GetRocket";
+import Score from "./Score";
+import PlayerGrid from "./PlayerGrid";
+import NPCGrid from "./NPCGrid";
 
 const Play = () => {
 
     return (
         <section>
-            <GetRocket />
+            {/* <GetRocket /> */}
             <Score />
             <PlayerGrid />
             <NPCGrid />
